+++ server/controllers/session.js
@@ -27,6 +27,7 @@ module.exports = {
         include: [{
           model: activity,
           as: 'activities',
+          include: [question]
         }],
       })
       .then(session => {